Replaced xvfb virtual display with headless Chrome
Chrome and Firefox now support headless mode by default so there's no more need of pyvirtualdisplay and xvfb.
I choose to use Chrome because for the moment on Ubuntu i have problems with headless firefox throwing "BrokenPipeError"  (https://github.com/SeleniumHQ/selenium/issues/6108) when trying to open urls. 

This is a first step toward the portability to Windows of the script.