Replaced Chrome with Firefox
With latest version of Selenium the previous issue (https://github.com/mozilla/geckodriver/issues/1321) that caused headless firefox browser to not work was solved, so there's no more need to stick to Chrome. 

I wanted to write a version of the script which worked with both Firefox and Chrome without the need of the user to comment/uncomment some sections of the source code, but it's really a mess to do something like this that would also be compatible on both linux and windows. 

I prefer readability and simplicity of the source-code in place of user-friendliness of the script itself.