Update to avoid annoying shadow ads to prevent clicking
Sometimes some "shadow" ads overlays all the webpage and prevents selenium clicking the buttons.
Generally on user click they open a popup window and disappear but on our case they cause a ElementClickInterceptedException and stop the script from running.
The workaround was to directly interact with the website events using javascript.